requirements.txt
black==25.1.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
imageio==2.37.0
imutils==0.5.3
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pyparsing==3.2.1
pytesseract==0.3.6
python-dateutil==2.9.0.post0
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
tifffile==2025.2.18
